🚚 api: move usecase to src/team context
