+++ api/src/team/domain/usecases/create-organization-invitation-by-admin.usecase.js
@@ -1,4 +1,4 @@
-import { OrganizationArchivedError } from '../../../src/shared/domain/errors.js';
+import { OrganizationArchivedError } from '../../../shared/domain/errors.js';
 
 const createOrganizationInvitationByAdmin = async function ({
   organizationId,
