+++ api/tests/team/unit/domain/usecases/create-organization-invitation-by-admin.usecase.test.js
@@ -1,7 +1,7 @@
-import { createOrganizationInvitationByAdmin } from '../../../../lib/domain/usecases/create-organization-invitation-by-admin.js';
-import { OrganizationArchivedError } from '../../../../src/shared/domain/errors.js';
-import { Membership } from '../../../../src/shared/domain/models/Membership.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { OrganizationArchivedError } from '../../../../../src/shared/domain/errors.js';
+import { Membership } from '../../../../../src/shared/domain/models/index.js';
+import { createOrganizationInvitationByAdmin } from '../../../../../src/team/domain/usecases/create-organization-invitation-by-admin.usecase.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | UseCase | create-organization-invitation-by-admin', function () {
   describe('#createOrganizationInvitationByAdmin', function () {
